hsakmt: Limit EoP queue entries to half queue entries
Limit number of entries that can fit in EoP queue to half the number of
entries that can fit in queue ring buffer. CP hardware does not hold all
bits for the read pointr, so we can run into situations when ring-buffer
wrap around happens and all packets in the queue are currently active.

Change-Id: I374bee94519bd25fd1d56977c045b629784916c8